fix Review

+++ src/main/java/com/sparta/projectblue/domain/review/service/ReviewService.java
@@ -3,7 +3,8 @@ package com.sparta.projectblue.domain.review.service;
 import com.sparta.projectblue.domain.performance.repository.PerformanceRepository;
 import com.sparta.projectblue.domain.reservation.entity.Reservation;
 import com.sparta.projectblue.domain.reservation.repository.ReservationRepository;
-import com.sparta.projectblue.domain.review.dto.ReviewRequestDto;
+import com.sparta.projectblue.domain.review.dto.CreateReviewDto;
+import com.sparta.projectblue.domain.review.dto.UpdateReviewDto;
 import com.sparta.projectblue.domain.review.entity.Review;
 import com.sparta.projectblue.domain.review.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,50 +17,55 @@ import org.springframework.transaction.annotation.Transactional;
 public class ReviewService {
 
     private final ReviewRepository reviewRepository;
-    private final PerformanceRepository performanceRepository;
+
     private final ReservationRepository reservationRepository;
 
     // 리뷰 생성
     @Transactional
-    public ReviewRequestDto.Response createReview(Long id, ReviewRequestDto.Request request) {
+    public CreateReviewDto.Response createReview(Long userId, CreateReviewDto.Request request) {
         Reservation reservation = reservationRepository.findById(request.getReservationId())
                 .orElseThrow(() -> new IllegalArgumentException("해당 예매를 찾을 수 없습니다."));
 
-        // performanceId는 reservation에서 가져옴
+        if (!reservation.getUserId().equals(userId)) {
+            throw new IllegalArgumentException("해당 예매의 사용자가 아닙니다");
+        }
+
         Review review = new Review(reservation.getPerformanceId(), reservation.getId(), request.getReviewRate(), request.getContents());
+
         reviewRepository.save(review);
-        return new ReviewRequestDto.Response(review, reservation.getPerformanceId());
-    }
 
-    // 예매 번호로 리뷰 조회
-    public ReviewRequestDto.Response getReview(Long reservationId) {
-
-        Review review = reviewRepository.findById(reservationId)
-//                    .orElseThrow(() -> new IllegalArgumentException("Payment not found for reservation"));
-                .orElse(null);
-
-        Reservation reservation = reservationRepository.findById(reservationId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 예매를 찾을 수 없습니다"));
-
-        return new ReviewRequestDto.Response(review, reservation.getPerformanceId());
+        return new CreateReviewDto.Response(review);
     }
 
     @Transactional
-    public ReviewRequestDto.Response updateReview(Long id, ReviewRequestDto.Request request) {
-        Reservation reservation = reservationRepository.findById(request.getReservationId())
-                .orElseThrow(() -> new IllegalArgumentException("해당 예매를 찾을 수 없습니다."));
-
+    public UpdateReviewDto.Response updateReview(Long userId, Long id, UpdateReviewDto.Request request) {
         Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 리뷰를 찾을 수 없습니다."));
 
-        review.updateReview(request.getReviewRate(), request.getContents());
-        reviewRepository.save(review);
+        Reservation reservation = reservationRepository.findById(review.getReservationId())
+                .orElseThrow(() -> new IllegalArgumentException("해당 예매를 찾을 수 없습니다."));
 
-        return new ReviewRequestDto.Response(review, reservation.getPerformanceId());
+        if (!reservation.getUserId().equals(userId)) {
+            throw new IllegalArgumentException("리뷰 작성자가 아닙니다");
+        }
+
+        review.updateReview(request.getReviewRate(), request.getContents());
+
+        return new UpdateReviewDto.Response(review);
     }
 
     @Transactional
-    public void deleteReview(Long id) {
+    public void deleteReview(Long userId, Long id) {
+        Review review = reviewRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 리뷰를 찾을 수 없습니다."));
+
+        Reservation reservation = reservationRepository.findById(review.getReservationId())
+                .orElseThrow(() -> new IllegalArgumentException("해당 예매를 찾을 수 없습니다."));
+
+        if (!reservation.getUserId().equals(userId)) {
+            throw new IllegalArgumentException("리뷰 작성자가 아닙니다");
+        }
+
         reviewRepository.deleteById(id);
     }
 }
